Obsolete products add

+++ API/Web/Service/ProductService.cs
@@ -11,6 +11,7 @@ using API.Web.Validators;
 using Microsoft.EntityFrameworkCore;
 using API.Web.Identity;
 using Web.Result.ErrorDefinitions;
+using System.Linq.Expressions;
 
 namespace API.Web.Service
 {
@@ -42,7 +43,7 @@ namespace API.Web.Service
         {
             try
             {
-                var result = _mapper.Map<IEnumerable<ProductDto>>(_productRepository.Find(x=>x.UserId == _userId || x.IsDefault == true));
+                var result = _mapper.Map<IEnumerable<ProductDto>>(_productRepository.Find(x=>((x.UserId == _userId || x.IsDefault == true) && x.IsAvailable)));
                 return new SuccessResult<IEnumerable<ProductDto>>(result);
             }
             catch(Exception ex)
@@ -55,17 +56,18 @@ namespace API.Web.Service
         public Result<ProductDto> GetProduct(int id)
         {
             try
-            {              
-                Product product;
-
+            {            
+                Expression<Func<Product,bool>> searchFunction;
                 if(!_isUserAdmin)
                 {
-                    product = _productRepository.Find(x=>x.UserId == _userId && x.Id == id).FirstOrDefault();
+                    searchFunction = x => x.UserId == _userId && x.Id == id && x.IsAvailable;
                 }
                 else
                 {
-                    product = _productRepository.Get(id);
+                    searchFunction = x => x.Id == id ;
                 }
+
+                var product = _productRepository.Find(searchFunction).FirstOrDefault();
 
                 if(product == null)
                 {
@@ -86,9 +88,19 @@ namespace API.Web.Service
         public Result<IEnumerable<ProductDto>> GetProducts(string productName)
         {
             try
-            { 
+            {
+                Expression<Func<Product,bool>> searchFunction;
+                if(!_isUserAdmin)
+                {
+                    searchFunction = x=> EF.Functions.Like(x.Name, $"%{productName}%") && x.UserId == _userId && x.IsAvailable;
+                }
+                else
+                {
+                    searchFunction = x=> EF.Functions.Like(x.Name, $"%{productName}%");
+                }
+
                 var products = _productRepository
-                    .Find(x=> EF.Functions.Like(x.Name, $"%{productName}%") && x.UserId == _userId)
+                    .Find(searchFunction)
                     .ToList();
 
                 var productsDto = _mapper.Map<IEnumerable<ProductDto>>(products);
@@ -119,10 +131,11 @@ namespace API.Web.Service
                 var productEntity = _mapper.Map<Product>(product);
 
                 productEntity.UserId = _userId;
-                
                 if(!_isUserAdmin)
                 {
+                    productEntity.IsAvailable = true;
                     productEntity.IsDefault = false;
+                    productEntity.DateAdded = DateTime.Now;
                 }
                 
                 var result = _productRepository.Add(productEntity);
@@ -136,7 +149,7 @@ namespace API.Web.Service
             }
         }
 
-        public Result<ProductDto> EditProduct(ProductDto product)
+        public Result<ProductDto> EditProduct(int id, ProductDto product)
         {
             try
             {
@@ -150,30 +163,39 @@ namespace API.Web.Service
 
                 if(!_isUserAdmin)
                 {
-                    productToEdit = _productRepository.Find(x=>x.UserId == _userId && x.Id == product.Id.Value).FirstOrDefault();
+                    productToEdit = _productRepository.Find(x=>x.UserId == _userId && x.Id == id && x.IsAvailable).FirstOrDefault();
+                    if(product.DateAdded == null)
+                    {
+                        product.DateAdded = DateTime.Now;
+                    }
+                    product.IsAvailable = true;
                     product.IsDefault = false;
+                    product.Id = null;
                 }
                 else
                 {
-                    productToEdit = _productRepository.Get(product.Id.Value);
+                    productToEdit = _productRepository.Get(id);
                 }
 
                 if(productToEdit == null)
                 {
-                    _logger.LogInformation($"Product with id = {product.Id} was not found!");
-                    return new NotFoundResult<ProductDto>(string.Format(ErrorDefinitions.NotFoundEntityWithIdError,new string[]{"Product",product.Id.ToString()}));
+                    _logger.LogInformation($"Product with id = {id} was not found!");
+                    return new NotFoundResult<ProductDto>(string.Format(ErrorDefinitions.NotFoundEntityWithIdError,new string[]{"Product",id.ToString()}));
                 }
 
-                if(!IsProductNameValid(_userId, product?.Id, product.Name))
+                if(!IsProductNameValid(_userId, id, product.Name))
                 {
                     return new InvalidResult<ProductDto>($"Product name {product.Name} is invalid!");
                 }
                                 
                 var productEntity = _mapper.Map<Product>(product);
-                var result = _productRepository.Update(productEntity);
+                productEntity.UserId = _userId;
+                productToEdit.IsAvailable = false;
+                var editResult = _productRepository.Update(productToEdit);
+                var addResult = _productRepository.Add(productEntity);
                 _productRepository.SaveChanges();
 
-                return new SuccessResult<ProductDto>(_mapper.Map<ProductDto>(result));
+                return new SuccessResult<ProductDto>(_mapper.Map<ProductDto>(addResult));
             }
             catch(Exception ex)
             {
@@ -186,15 +208,17 @@ namespace API.Web.Service
         {
             try
             {
-                Product productToDelete;
+                Expression<Func<Product,bool>> searchFunction;
                 if(!_isUserAdmin)
                 {
-                    productToDelete = _productRepository.Find(x=>x.UserId == _userId && x.Id == id).FirstOrDefault();
+                    searchFunction = x => x.UserId == _userId && x.Id == id && x.IsAvailable;
                 }
                 else
                 {
-                    productToDelete = _productRepository.Get(id);
+                    searchFunction = x => x.Id == id ;
                 }
+
+                var productToDelete = _productRepository.Find(searchFunction).FirstOrDefault();
 
                 if(productToDelete == null)
                 {
@@ -202,7 +226,19 @@ namespace API.Web.Service
                     return new NotFoundResult<ProductDto>(string.Format(ErrorDefinitions.NotFoundEntityWithIdError,new string[]{"Product",id.ToString()}));
                 }
 
-                var result = _productRepository.Delete(productToDelete);
+                Product result;
+
+                if(!_isUserAdmin)
+                {
+                    productToDelete.IsAvailable = false;
+                    result = _productRepository.Update(productToDelete);
+                }
+                else
+                {
+                    productToDelete = _productRepository.Get(id);
+                    result = _productRepository.Delete(productToDelete);
+                }
+
                 _productRepository.SaveChanges();
                 return new SuccessResult<ProductDto>(_mapper.Map<ProductDto>(result));
             }
@@ -229,7 +265,7 @@ namespace API.Web.Service
 
         private bool IsProductNameValid(int userId, int? productId, string productName)
         {
-            if(_productRepository.Find(x=>x.Name == productName && (x.UserId == userId || x.IsDefault) && (productId.HasValue ? x.Id != productId : true)).FirstOrDefault() != null)
+            if(_productRepository.Find(x=>x.Name == productName && x.IsAvailable && (x.UserId == userId || x.IsDefault) && (productId.HasValue ? x.Id != productId : true)).FirstOrDefault() != null)
             {
                 return false;
             }
